requirements.txt
yt-dlp
jinja2
python-dotenv
rich
requests
